# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: projects
    description: Project operations
  - name: files
    description: File operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
        "409":
          description: User already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
        "401":
          description: Invalid credentials
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: Get all projects for authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a specific project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project with files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWithFiles"
        "401":
          description: Unauthorized
        "404":
          description: Project not found
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Project not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project deleted
        "401":
          description: Unauthorized
        "404":
          description: Project not found
  /files/project/{projectId}:
    get:
      operationId: getProjectFiles
      tags: [files]
      summary: Get all files for a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File"
        "401":
          description: Unauthorized
        "404":
          description: Project not found
    post:
      operationId: createFile
      tags: [files]
      summary: Create a new file in a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileRequest"
      responses:
        "201":
          description: File created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Project not found
        "409":
          description: File already exists
  /files/project/{projectId}/bulk:
    post:
      operationId: bulkCreateFiles
      tags: [files]
      summary: Bulk create/update files for a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateFilesRequest"
      responses:
        "201":
          description: Files created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Project not found
  /files/{id}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get a specific file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: File not found
    put:
      operationId: updateFile
      tags: [files]
      summary: Update a file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFileRequest"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: File not found
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: File not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 8
      required:
        - email
        - username
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
        token:
          type: string
      required:
        - userId
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - createdAt
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        tree:
          type: string
      required:
        - name
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        tree:
          type: string
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        tree:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - tree
        - createdAt
        - updatedAt
    ProjectWithFiles:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            files:
              type: array
              items:
                $ref: "#/components/schemas/File"
    CreateFileRequest:
      type: object
      properties:
        path:
          type: string
          minLength: 1
        content:
          type: string
      required:
        - path
    UpdateFileRequest:
      type: object
      properties:
        path:
          type: string
          minLength: 1
        content:
          type: string
    File:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        path:
          type: string
        content:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - path
        - content
        - updatedAt
    BulkCreateFilesRequest:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/CreateFileRequest"
      required:
        - files